Favorites: added DB, added JS to add/remove a favorite, added buttons to show page.
The base favorites functionality has been created.

-The Boomarks Tabl
e is created on the initial DB load (just an empty bookmark table with a
n autoincrement id and a show id).
-Added buttons to the Show page so
that you can add / remove a favorite.

todo:
-Some wonkiness on whether
 we show the blue/black button.  The text is always correct but for some
 reason the button color is wrong.  Looking into it.
-Still haven't made
 the "favorites" page itself, but that will come next.
-Will refactor th
e Favorite JS function so that it decouples the UI from the Add/Remove l
ogic (so that other pages can call it)