Empty favorite view done (issue #32) - comments very welcom
I tried to mimic the iPhone empty screen as much as possible.  I think
we might need to tweak the star image a bit (or some of the grays) but
this should be functioning pretty well right now (and provide a much
better prompt to the user than a blank screen).  please let me know what
you guys think: I'm not the best at UI, but I can take notes well.